refactor: jdbc 미션 리팩토링
