+++ src/main/java/roomescape/RenderController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
