+++ src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.reservation.domain;
 
 public class Reservation {
