+++ src/main/java/roomescape/reservation/persistence/ReservationJdbcRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.reservation.persistence;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -11,7 +11,7 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 
-import roomescape.domain.Reservation;
+import roomescape.reservation.domain.Reservation;
 
 @Repository
 @Primary
