+++ src/main/java/roomescape/reservation/presentation/ReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.reservation.presentation;
 
 import java.net.URI;
 import java.util.List;
@@ -9,12 +9,14 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import roomescape.dto.request.ReservationSaveRequest;
-import roomescape.dto.response.ReservationResponse;
-import roomescape.service.ReservationService;
+import roomescape.reservation.application.ReservationService;
+import roomescape.reservation.presentation.dto.request.ReservationSaveRequest;
+import roomescape.reservation.presentation.dto.response.ReservationResponse;
 
+@RequestMapping("/reservations")
 @RestController
 public class ReservationController {
 
@@ -24,19 +26,19 @@ public class ReservationController {
 		this.reservationService = reservationService;
 	}
 
-	@GetMapping("/reservations")
+	@GetMapping
 	public List<ReservationResponse> getReservations() {
 		return reservationService.getReservations();
 	}
 
-	@PostMapping("/reservations")
+	@PostMapping
 	public ResponseEntity<ReservationResponse> saveReservation(@RequestBody ReservationSaveRequest request) {
 		ReservationResponse response = reservationService.saveReservation(request);
 		URI location = URI.create("/reservations/" + response.id());
 		return ResponseEntity.created(location).body(response);
 	}
 
-	@DeleteMapping("/reservations/{id}")
+	@DeleteMapping("/{id}")
 	public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
 		reservationService.deleteReservation(id);
 		return ResponseEntity.noContent().build();
