+++ src/main/java/roomescape/reservation/presentation/ReservationExceptionHandler.java
@@ -1,11 +1,11 @@
-package roomescape.controller;
+package roomescape.reservation.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 
-import roomescape.exception.BadReservationSaveRequestException;
-import roomescape.exception.NotFoundReservationException;
+import roomescape.reservation.presentation.exception.BadReservationSaveRequestException;
+import roomescape.reservation.presentation.exception.NotFoundReservationException;
 
 @ControllerAdvice
 public class ReservationExceptionHandler {
